--- backend/Dockerfile
@@ -1,11 +1,0 @@
-FROM python:3.9
-
-WORKDIR /backend
-
-COPY . .
-
-RUN pip install -r requirements.txt
-
-EXPOSE 5000
-
-CMD gunicorn --bind 0.0.0.0:5000 api:app
