+++ backend/sell/Dockerfile
@@ -2,7 +2,8 @@ FROM python:3.9
 
 WORKDIR /backend/sell
 
-COPY . .
+COPY ./sell/sellservice.py .
+COPY ../requirements.txt .
 
 RUN pip install -r requirements.txt
