+++ docker-compose.yml
@@ -6,9 +6,17 @@ services:
     ports:
       - "27017:27017"
 
-  backend:
+  backend_buy:
     build:
       context: ./backend
+      dockerfile: ./buy/Dockerfile
+    ports:
+      - "5000"
+
+  backend_sell:
+    build:
+      context: ./backend
+      dockerfile: ./sell/Dockerfile
     ports:
       - "5000"
 
@@ -24,6 +32,8 @@ services:
       - ./nginx.conf:/etc/nginx/nginx.conf:ro
     depends_on:
       - quoteserver
-      - backend
+      - backend_buy
+      - backend_sell
+
     ports:
       - "80:80"
