+++ nginx.conf
@@ -3,14 +3,40 @@ events {
 }
 
 http {
+
+    upstream backend_buy {
+        server backend_buy:5000;
+    }
+
+    upstream backend_sell {
+        server backend_sell:5000;
+    }
+
+    # upstream backend_user {
+    #     server backend_user:5000;
+    # }
+
     server {
         listen 80;
 
-        location / {
-            proxy_pass http://backend:5000;
+        location /buy/ {
+            proxy_pass http://backend_buy/;
             proxy_set_header Host $host;
             proxy_set_header X-Real-IP $remote_addr;
         }
+
+        location /sell/ {
+            proxy_pass http://backend_sell/;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+        }
+
+
+        # location / {
+        #     proxy_pass http://backend:5000/;
+        # }
+
     }
+
 }
